+++ apps/demo-react/components/info/styles.tsx
@@ -1,9 +1,0 @@
-import styled from 'styled-components';
-
-export const Line = styled.div`
-  margin: 8px 0;
-`;
-
-export const Heading = styled.h4`
-  margin: 12px 0;
-`;