+++ apps/demo-react/global.d.ts
@@ -1,0 +1,12 @@
+declare module '*.svg' {
+  /**
+   * Use `any` to avoid conflicts with
+   * `@svgr/webpack` plugin or
+   * `babel-plugin-inline-react-svg` plugin.
+   */
+  const content: any;
+  export const ReactComponent: React.FunctionComponent<
+    React.ComponentProps<'svg'>
+  >;
+  export default content;
+}