+++ apps/demo-react/next.config.js
@@ -1,22 +1,33 @@
+import buildDynamics from './scripts/build-dynamics.mjs';
 import NextBundleAnalyzer from '@next/bundle-analyzer';
 
-const alchemyApiKey = process.env.ALCHEMY_API_KEY;
-const analyzeBundle = process.env.ANALYZE_BUNDLE ?? false;
+buildDynamics();
+
+const ANALYZE_BUNDLE = process.env.ANALYZE_BUNDLE === 'true';
+const basePath = process.env.BASE_PATH || '';
 
 const withBundleAnalyzer = NextBundleAnalyzer({
-  enabled: analyzeBundle,
+  enabled: ANALYZE_BUNDLE,
 });
 
 export default withBundleAnalyzer({
   reactStrictMode: true,
-  basePath: process.env.BASE_PATH || '',
+  basePath,
   compiler: {
     styledComponents: {
-      ssr: true
-    }
+      ssr: true,
+    },
   },
-  publicRuntimeConfig: {
-    alchemyApiKey,
+  webpack(config) {
+    config.module.rules.push(
+      // Teach webpack to import svg and md files
+      {
+        test: /\.svg$/,
+        use: ['@svgr/webpack', 'url-loader'],
+      },
+    );
+
+    return config;
   },
   async headers() {
     return [
@@ -25,19 +36,22 @@ export default withBundleAnalyzer({
         source: '/manifest.json',
         headers: [
           {
-            key: "Access-Control-Allow-Origin",
+            key: 'Access-Control-Allow-Origin',
             value: '*',
           },
           {
-            key: "Access-Control-Allow-Methods",
+            key: 'Access-Control-Allow-Methods',
             value: 'GET',
           },
           {
-            key: "Access-Control-Allow-Headers",
+            key: 'Access-Control-Allow-Headers',
             value: 'X-Requested-With, content-type, Authorization',
           },
         ],
       },
-    ]
-  }
-})
+    ];
+  },
+  serverRuntimeConfig: {
+    basePath,
+  },
+});